Reorganize modules and import tests from launch_ros PR
